// src/features/add-measurement/model/types.ts
/**
 * Типы для фичи добавления/редактирования замера
 */

export interface MeasurementFormData {
  date: string;
  weight: string;
  biceps: string;
  forearm: string;
  shoulders: string;
  chest: string;
  waist: string;
  glutes: string;
  hips: string;
  calves: string;
}